refactor: fieldState.ts를 components에서 hooks 폴더로 이동
