+++ src/feature/CardRegister/hooks/useCvcNumber.ts
@@ -1,5 +1,5 @@
 import {useState} from 'react';
-import {createFlags, computeNextTouched, computeNextErrorInfo} from '../components/InfoInputSection/fieldState';
+import {createFlags, computeNextTouched, computeNextErrorInfo} from './fieldState';
 
 const CVC_LENGTH = 3;
 const ERROR_MSG = 'CVC 번호 3자리를 입력해 주세요';
